+++ packages/uni_app/lib/view/faculty/faculty_old.dart
@@ -6,15 +6,14 @@ import 'package:uni/utils/navigation_items.dart';
 import 'package:uni/view/calendar/widgets/calendar_card.dart';
 import 'package:uni/view/common_widgets/generic_expansion_card.dart';
 import 'package:uni/view/common_widgets/pages_layouts/general/general.dart';
-import 'package:uni/view/faculty/widgets/academic_services_card.dart';
+import 'package:uni/view/faculty/widgets/academic_services_card_old.dart';
 import 'package:uni/view/faculty/widgets/copy_center_card.dart';
 import 'package:uni/view/faculty/widgets/dona_bia_card.dart';
 import 'package:uni/view/faculty/widgets/infodesk_card.dart';
 import 'package:uni/view/faculty/widgets/multimedia_center_card.dart';
 import 'package:uni/view/faculty/widgets/other_links_card.dart';
 import 'package:uni/view/faculty/widgets/sigarra_links_card.dart';
-//import 'package:uni/view/library/widgets/library_occupation_card.dart';
-import 'package:uni_ui/cards/library_occupation_card.dart';
+import 'package:uni/view/library/widgets/library_occupation_card.dart';
 
 class FacultyPageView extends StatefulWidget {
   const FacultyPageView({super.key});
@@ -32,9 +31,8 @@ class FacultyPageViewState extends GeneralPageViewState {
   Widget getBody(BuildContext context) {
     return ListView(
       children: [
-        //LibraryOccupationCard(),
+        LibraryOccupationCard(),
         CalendarCard(),
-        AcademicServicesCard(),
         ...getUtilsSection(),
       ],
     );
@@ -42,13 +40,13 @@ class FacultyPageViewState extends GeneralPageViewState {
 
   List<GenericExpansionCard> getUtilsSection() {
     return const [
-      //AcademicServicesCard(),
-      /*InfoDeskCard(),
+      AcademicServicesCard(),
+      InfoDeskCard(),
       DonaBiaCard(),
       CopyCenterCard(),
       MultimediaCenterCard(),
       SigarraLinksCard(),
-      OtherLinksCard(),*/
+      OtherLinksCard(),
     ];
   }
 
@@ -57,7 +55,4 @@ class FacultyPageViewState extends GeneralPageViewState {
     return Provider.of<LibraryOccupationProvider>(context, listen: false)
         .forceRefresh(context);
   }
-
-
-
 }
